\name{com}
\alias{com}
\docType{data}
\title{
com
}
\description{
A simple binary community matrix.
}
\usage{data("com")}
\format{
  The format is:
 num [1:5, 1:10] 0 1 0 0 0 0 0 0 0 1 ...
 - attr(*, "dimnames")=List of 2
  ..$ : chr [1:5] "Sample_1" "Sample_2" "Sample_3" "Sample_4" ...
  ..$ : chr [1:10] "Sp_1" "Sp_2" "Sp_3" "Sp_4" ...
}
\details{
5 samples in rows and 10 taxa in columns.
}
\keyword{datasets}
