% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist2vec.R
\name{dist2vec}
\alias{dist2vec}
\title{Distance to Vector}
\usage{
dist2vec(d)
}
\arguments{
\item{d}{A distance matrix.}
}
\value{
A vector of the distances.
}
\description{
This function converts a distance object into a vector. The vectors may then
  be incorporated into a summary table.
}
\details{
Converts a distance object into a vector.
}
\examples{
data(dune)
d <- vegdist(dune, method = "bray")
dist2vec(d)

}
\author{
John Quensen
}
