\name{dune}
\alias{dune}
\docType{data}
\title{
dune
}
\description{
Community matrix dune from the vegan package.
}
\usage{data("dune")}
\format{
  A data frame with 20 observations on the following 30 variables.
  \describe{
    \item{\code{Achimill}}{a numeric vector}
    \item{\code{Agrostol}}{a numeric vector}
    \item{\code{Airaprae}}{a numeric vector}
    \item{\code{Alopgeni}}{a numeric vector}
    \item{\code{Anthodor}}{a numeric vector}
    \item{\code{Bellpere}}{a numeric vector}
    \item{\code{Bromhord}}{a numeric vector}
    \item{\code{Chenalbu}}{a numeric vector}
    \item{\code{Cirsarve}}{a numeric vector}
    \item{\code{Comapalu}}{a numeric vector}
    \item{\code{Eleopalu}}{a numeric vector}
    \item{\code{Elymrepe}}{a numeric vector}
    \item{\code{Empenigr}}{a numeric vector}
    \item{\code{Hyporadi}}{a numeric vector}
    \item{\code{Juncarti}}{a numeric vector}
    \item{\code{Juncbufo}}{a numeric vector}
    \item{\code{Lolipere}}{a numeric vector}
    \item{\code{Planlanc}}{a numeric vector}
    \item{\code{Poaprat}}{a numeric vector}
    \item{\code{Poatriv}}{a numeric vector}
    \item{\code{Ranuflam}}{a numeric vector}
    \item{\code{Rumeacet}}{a numeric vector}
    \item{\code{Sagiproc}}{a numeric vector}
    \item{\code{Salirepe}}{a numeric vector}
    \item{\code{Scorautu}}{a numeric vector}
    \item{\code{Trifprat}}{a numeric vector}
    \item{\code{Trifrepe}}{a numeric vector}
    \item{\code{Vicilath}}{a numeric vector}
    \item{\code{Bracruta}}{a numeric vector}
    \item{\code{Callcusp}}{a numeric vector}
  }
}
\details{
A community matrix with samples as rows and taxa as columns.
}
\source{
Copied from the vegan package.
}
\references{
  Jari Oksanen, F. Guillaume Blanchet, Michael Friendly, Roeland Kindt, Pierre
  Legendre, Dan McGlinn, Peter R. Minchin, R. B. O'Hara, Gavin L. Simpson, Peter
  Solymos, M. Henry H. Stevens, Eduard Szoecs and Helene Wagner (2019). vegan:
  Community Ecology Package. R package version 2.5-6.
  https://CRAN.R-project.org/package=vegan
}
\examples{
data(dune)
## maybe str(dune) ; plot(dune) ...
}
\keyword{datasets}
