% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_oecosimu_dist.R
\name{format_oecosimu_dist}
\alias{format_oecosimu_dist}
\title{Reformat an oecosimu Object}
\usage{
format_oecosimu_dist(sol, com)
}
\arguments{
\item{sol}{Result of oecosimu when statistic is a matrix.}

\item{com}{Community matrix used to calculate oecosimu result.}
}
\value{
A data frame.
}
\description{
Reformats an oecosim object if the statistic returned is a distance matrix.
}
\details{
Vegan's oecosimu function is used to compare an observed and expected value or
  values based on a null model. A drawback of the oecosimu summary output is that sample 
  pairs are not named.  
  
  This function reformats the oecosimu result into a table giving the names of sample
  pairs, the observed statistic, the expected statistic and its standard deviation based 
  on the null model, the standard effect size, the selection strength as defined in Zhou 
  et al. (2014), and the probability that the observed and expected values of the statistic
  are not different. This function is appropriate when the statistic is a distance matrix.
}
\references{
Zhou JZ, Deng Y, Zhang P et al. (2014) Stochasticity, succession, and environmental
  perturbations in a fluidic ecosystem. Proceedings of the National Academy of Sciences of the
  United States of America, 111, E836-E845.
}
\seealso{
format_oecosimu_vector
}
\author{
John Quensen
}
