% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix2list.R
\name{matrix2list}
\alias{matrix2list}
\title{Convert a square matrix to a list}
\usage{
matrix2list(x)
}
\arguments{
\item{x}{A square matrix.}
}
\value{
A vector.
}
\description{
This function converts a square matrix as returned by a null model calculation for a similarity or distance into a list.
}
\author{
John Quensen
}
