% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_shared_spp.R
\name{mean_shared_spp}
\alias{mean_shared_spp}
\title{Mean Shared Species}
\usage{
mean_shared_spp(com)
}
\arguments{
\item{com}{A community matrix with samples as rows and species as columns.}
}
\value{
A real number.
}
\description{
Calculates the mean number of shared species from a community matrix for all sample pairs.
}
\details{
If presented with a community matrix of count data, this function first converts 
  it to presence/absence.
}
\examples{
data(com)
mean_shared_spp(com)
}
\author{
John Quensen
}
