% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perm_disper.R
\name{perm_disper}
\alias{perm_disper}
\title{Permutational Dispersion Analysis}
\usage{
perm_disper(com, w, n.reps=999)
}
\arguments{
\item{com}{A binary community matrix with samples in rows and taxa in columns.}

\item{w}{Sampling weights.}

\item{n.reps}{Number of replications.}
}
\value{
Result of vegan's betadisper test between observed and null model expectation.
}
\description{
This function tests if the dispersion about a group centroid is significantly different
  from null model expectation. It makes the test using Jaccard distances and randomizes 
  the matrix using the randomize_by_row function. It is based on the procedure used in  
  Zhou et al. (2014).
}
\details{
The binary community matrix is randomized using the function randomize_by_rows. 
  Sampling weights are species sums by default, but may be specified differently by the 
  user. This method maintains sample sums (species richness) and constrains species 
  occurrences.
}
\references{
Zhou, J. Z., Y. Deng, P. Zhang, K. Xue, Y. T. Liang, J. D. Van Nostrand, 
  Y. F. Yang, Z. L. He, L. Y. Wu, D. A. Stahl, T. C. Hazen, J. M. Tiedje, and A. P. 
  Arkin. 2014. Stochasticity, succession, and environmental perturbations in a fluidic 
  ecosystem. Proceedings of the National Academy of Sciences of the United States of 
  America 111:E836-E845.
 @importFrom vegan vegdist, betadisper
}
\author{
John Quensen
}
