% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomize_by_rows.R
\name{randomize_by_rows}
\alias{randomize_by_rows}
\title{Randomize by Rows}
\usage{
randomize_by_rows(com, w, rename=FALSE)
}
\arguments{
\item{com}{A binary community matrix with samples as rows and taxaa as columns.}

\item{w}{Sampling weights in a vector of length equal to row length in com.}

\item{rename}{A logical indicating whether or not samples should be renamed.}
}
\value{
A randomized binary matrix.
}
\description{
This function randomizes a binary matrix by shuffling the values in rows using sampling 
  probabilities specified in vector w. It is similar to Sim4 but has the advantage that 
  the user may specify w instead of having it calculated from the column sums.
}
\details{
The number of non-zero entries in w must be at least the maximum number of 
  non-zero entries in a row of com (max(rowSums(com)).
}
\author{
John Quensen
}
