% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raup_crick.R
\name{raup_crick}
\alias{raup_crick}
\title{Chase's Raup-Crick}
\usage{
raup_crick(spXsite, plot_names_in_col1 = FALSE, classic_metric = FALSE, 
  split_ties = TRUE, reps = 999, set_all_species_equal = FALSE, 
  as.distance.matrix = TRUE, report_similarity = FALSE)
}
\arguments{
\item{spXsite}{A community matrix with samples as rows and species as columns.}

\item{plot_names_in_col1}{A logical specifying whether or not the sample names 
are in the first column of the community matrix.}

\item{classic_metric}{A logical specifying whether or not the Raup-Crick metric 
should be standardized to the range -1 to 1 as Chase did.}

\item{split_ties}{A logical specifying how ties are handled in calculating probabilities.}

\item{reps}{The number of replications used in calculating null expectations.}

\item{set_all_species_equal}{A logical specifying sampling probablilites.}

\item{as.distance.matrix}{A logical specifying whether or not to return the result 
as a distance matrix.}

\item{report_similarity}{A logical specifying whether or not to return the result as similarities.}
}
\value{
A distance matrix or similarity matrix, depending on parameters.
}
\description{
Calculates a modified version of the Raup-Crick metric using Chase's original R script 
  included in the supplementary material to Chase et al. (2011).
}
\details{
This function calculates a modified version of Raup-Crick using a null model approach.
By default, it standardizes the metric to range from -1 to 1 instead of 0 to 1. When this is 
done, values near -1 indicate that samples are more similar than expected by chance and values
near 1 indicate that samples are less similar than expected by chance. If the result is to be 
used as a distance method, the metric should not be re-scaled (set classic_metric to TRUE).  

The argument report_similarity defaults to FALSE so the function reports a dissimilarity 
(which is appropriate as a measure of beta diversity).  Setting report_similarity=TRUE returns
a measure of similarity, as Raup and Crick originally specified.  

If ties are split (as Chase recommends) the dissimilarity (default) and similarity (set 
report_similarity=TRUE) calculations can be flipped by multiplying by -1 (for Chase's 
modification, which ranges from -1 to 1) or by subtracting the metric from 1 (for the classic 
metric which ranges from 0 to 1). If ties are not split (and there are ties between the observed
and expected shared number of species) this conversion will not work. 

Raup-Crick is calculated from a presence/absence communitiy matrix. If this function is 
given count data, it first converts it to presence/absence.  

The choice of how many plots (rows) to include has a real impact  on the metric, as species
and their occurrence frequencies across the set of plots is used to determine gamma and the
frequency with which each species is drawn from the null model.
}
\examples{
data(com)
dist.rc <- raup_crick(com, classic_metric=TRUE)
}
\references{
Chase, J. M., N. J. B. Kraft, K. G. Smith, M. Vellend, and B. D. Inouye. 
  2011. Using null models to disentangle variation in community dissimilarity from 
  variation in alpha-diversity. Ecosphere 2(2): Article 24.  
  
  Raup, D. M. and R. E. Crick. 1979. Measurement of faunal similarity in paleontology.
  Journal of Paleontology 53:1213-1227.
}
\author{
Jonathan M. Chase; help edited by John Quensen
}
