% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_spp.R
\name{shared_spp}
\alias{shared_spp}
\title{Shared Species}
\usage{
shared_spp(com)
}
\arguments{
\item{com}{A community matrix with samples as rows and species as columns.}
}
\value{
A vector of the number of shared species..
}
\description{
Calculates shared species between all samples in a presenece/absence community matrix.
}
\details{
If presented with a community matrix of count data, this function first 
  converts it to presence/absence.
}
\examples{
data(com)
shared_spp(com)
}
\author{
John Quensen
}
