% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim10.R
\name{Sim10}
\alias{Sim10}
\title{Gotelli's Sim10 Function}
\usage{
Sim10(m, Row.Weights, Col.Weights)
}
\arguments{
\item{m}{A binary community matrix.}

\item{Row.Weights}{User supplied sampling weights for rows.}

\item{Col.Weights}{User supplied sampling weights for columns.}
}
\value{
A randomized binary matrix of the same dimensions as the input matrix.
}
\description{
Randomizes a binary matrix m by reshuffling all elements. Rows & columns proportional
  to supplied row and column weights.
}
\details{
Makes a call to the function VectorSample.
}
\examples{
data(com)
rw <- sample(1:10, nrow(com), replace=TRUE)
cw <- sample(1:10, ncol(com), replace=TRUE)
rand.com <- Sim10(com, Row.Weights=rw, Col.Weights=cw)
}
\references{
Gotelli, N. J. 2000. Null model analysis of species co-occurrence patterns.
  Ecology 81:2606-2621.
}
\author{
Nicholas J. Gotelli. Help edited by John Quensen.
}
