% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim2.R
\name{Sim2}
\alias{Sim2}
\title{Gotelli's Sim2 Function}
\usage{
Sim2(m)
}
\arguments{
\item{m}{A binary community matrix.}
}
\value{
A binary matrix randomized by rows of the same dimensions as the input matrix.
}
\description{
Randomizes a binary matrix m by reshuffling elements within each row equiprobably.
}
\examples{
m <- sample(c(1,0), 50, replace=TRUE)
m <- matrix(m, 5, 10)
random.m <- Sim2(m)
}
\references{
Gotelli, N. J. 2000. Null model analysis of species co-occurrence patterns. Ecology 81:2606-2621.
}
\author{
Nicholas J. Gotelli. Help edited by John Quensen.
}
