% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim4.R
\name{Sim4}
\alias{Sim4}
\title{Gotelli's Sim4 Function}
\usage{
Sim4(m)
}
\arguments{
\item{m}{A binary community matrix.}
}
\value{
A binary matrix randomized by columns and of the same dimensions as the
  input matrix.
}
\description{
Randomizes a binary matrix m by shuffling elements within each row. Sampling weights
  for each column are proportional to column sums.
}
\details{
Makes a call to VectorSample.
}
\examples{
m <- sample(c(1,0), 50, replace=TRUE)
m <- matrix(m, 5, 10)
random.m <- Sim4(m)
}
\references{
Gotelli, N. J. 2000. Null model analysis of species co-occurrence 
  patterns. Ecology 81:2606-2621.
}
\author{
Nicholas J. Gotelli. Help edited by John Quensen.
}
