% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim5.R
\name{Sim5}
\alias{Sim5}
\title{Gotelli's Sim5 Function}
\usage{
Sim5(m)
}
\arguments{
\item{m}{A binary community matrix.}
}
\value{
A binary matrix randomized by columns of the same dimensions as the input matrix.
}
\description{
Randomizes a matrix m by reshuffling elements within each column. Sampling weights
  for each row are proportional to row sums.
}
\examples{
m <- sample(c(1,0), 50, replace=TRUE)
m <- matrix(m, 5, 10)
random.m <- Sim5(m)
}
\references{
Gotelli, N. J. 2000. Null model analysis of species co-occurrence patterns.
  Ecology 81:2606-2621.
}
\author{
Nicholas J. Gotelli. Help edited by John Quensen.
}
