% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim6.R
\name{Sim6}
\alias{Sim6}
\title{Gotelli's Sim6 Function}
\usage{
Sim6(m)
}
\arguments{
\item{m}{A binary community matrix.}
}
\value{
A randomized binary matrix of the same dimensions as the input matrix.
}
\description{
Randomizes a binary matrix m by reshuffling all elements. Rows are equiprobable, 
  columns proportional to column sums.
}
\examples{
m <- sample(c(1,0), 50, replace=TRUE)
m <- matrix(m, 5, 10)
random.sam <- Sim6(m) 
}
\references{
Gotelli, N. J. 2000. Null model analysis of species co-occurrence patterns.
  Ecology 81:2606-2621.
}
\author{
Nicholas J. Gotelli. Help edited by John Quensen.
}
