% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim7.R
\name{Sim7}
\alias{Sim7}
\title{Gotelli's Sim7 Function}
\usage{
Sim7(m)
}
\arguments{
\item{m}{A binary matrix.}
}
\value{
A randomized binary matrix.
}
\description{
Randomizes a binary matrix m by reshuffling all elements. Columns are equiprobable,
  rows proportional to row sums.
}
\examples{
m <- sample(c(1,0), 50, replace=TRUE)
m <- matrix(m, 5, 10)
random.sam <- Sim7(m)

}
\references{
Gotelli, N. J. 2000. Null model analysis of species co-occurrence patterns.
  Ecology 81:2606-2621.
}
\author{
Nicholas J. Gotelli. Help edited by John Quensen.
}
