% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim8.R
\name{Sim8}
\alias{Sim8}
\title{Sim8 Function}
\usage{
Sim8(m)
}
\arguments{
\item{m}{A binary community matrix.}
}
\value{
A randomized binary matrix of the same dimensions as the input matrix.
}
\description{
Randomizes a binary matrix m by reshuffling all elements. Columns are proportional
  to column sums, rows proportional to row sums.
}
\details{
Makes a call to VectorSample.
}
\examples{
m <- sample(c(1,0), 50, replace=TRUE)
m <- matrix(m, 5, 10)
random.sam <- Sim8(m)
}
\references{
Gotelli, N. J. 2000. Null model analysis of species co-occurrence patterns.
  Ecology 81:2606-2621.

Gotelli, N. J. 2000. Null model analysis of species co-occurrence patterns.
  Ecology 81:2606-2621.
}
\author{
Nicholas J. Gotelli. Help edited by John Quensen.
}
