% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simberloff_sum.R
\name{simberloff_sum}
\alias{simberloff_sum}
\title{Simberloff Summary}
\usage{
simberloff_sum(rslt.sum, alpha = 0.05)
}
\arguments{
\item{rslt.sum}{Result from function format_oecosimu_vector.}

\item{alpha}{Confidence level.}
}
\value{
A data frame.
}
\description{
Makes a table of the number of times observed shared species are greater or lesser
  than expected, and number of times significantly so.
}
\references{
Simberloff, D. 1978. Using island biogeographic distributions to determine
  if colonizatioon is stochastic. American Naturalist 112:713-726.
}
\seealso{
format_oecosimu_vector
}
\author{
John Quensen
}
