% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat2dist.R
\name{stat2dist}
\alias{stat2dist}
\title{Convert Vector to Distance}
\usage{
stat2dist(com, stat, diagonal = TRUE)
}
\arguments{
\item{com}{The community matrix for which the vector of distances was calculated.}

\item{stat}{One of the statistics from the ecosimu function: statistic, mean.}

\item{diagonal}{A logical determining if distance matrix is square or without upper  half and diagonal.}
}
\value{
A distance matrix.
}
\description{
This function converts a vector of distances into a distance matrix.
}
\details{
The oecosimu function in vegan may be used to calculate the observed distances
  for a community matrix and the expected distances based on a null model, but the output
  is in the form of vectors. This function converts such vectors into a distance object 
  that may be used for ordination or clustering of samples. In the oecosimu output, 
  statistic is the observed value and mean is the expected value based on the null model.
}
\author{
John Quensen
}
