% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochasticity_is.R
\name{stochasticity_is}
\alias{stochasticity_is}
\title{Stochasticity by Independent Swap}
\usage{
stochasticity_is(com, n.reps=999)
}
\arguments{
\item{com}{A binary community matrix.}

\item{n.reps}{Number of replications.}
}
\value{
A vector of stochasticity values.
}
\description{
This function calculates stochasticity based on the Jaccard dissimilarity as defined
  in Zhou et al. (2014) by randomizing a binary community matrix with picante's 
  independentswap algorithm.
}
\details{
The binary community matrix is randomized using picante's function 
  randomizeMatrix with method="independentswap." The independent swap algorithm 
  maintains both row and column totals.
}
\references{
Zhou, J. Z., Y. Deng, P. Zhang, K. Xue, Y. T. Liang, J. D. Van Nostrand,
  Y. F. Yang, Z. L. He, L. Y. Wu, D. A. Stahl, T. C. Hazen, J. M. Tiedje, and 
  A. P. Arkin. 2014. Stochasticity, succession, and environmental perturbations 
  in a fluidic ecosystem. Proceedings of the National Academy of Sciences of the 
  United States of America 111:E836-E845.
}
\author{
John Quensen
}
