% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochasticity_rbr.R
\name{stochasticity_rbr}
\alias{stochasticity_rbr}
\title{Stochasticity by randomize_by_rows}
\usage{
stochasticity_rbr(com, w, n.reps=999)
}
\arguments{
\item{com}{A binary community matrix.}

\item{w}{A vector of sampling wieghts.}

\item{n.reps}{Number of replications.}
}
\value{
A vector of stochasticity values.
}
\description{
This function calculates stochasticity based on the Jaccard dissimilarity as defined
  in Zhou et al. (2014) by randomizing a binary community matrix using the function 
  randomize_by_rows in this package.
}
\details{
The binary community matrix is randomized using the function randomize_by_rows
  in this package. This method maintains row (sample) totals.
  
  The user must supply w, a vector of sampling weights. If w equals the column sums of
  com, the method corresponds to Ye Deng's gamma.method=sample (see reference). If w 
  equals the column sums of a larger experiment matrix of which com is a sub-set, the
  method corresponds to gamma.method=total.
}
\references{
Zhou, J. Z., Y. Deng, P. Zhang, K. Xue, Y. T. Liang, J. D. Van Nostrand, 
  Y. F. Yang, Z. L. He, L. Y. Wu, D. A. Stahl, T. C. Hazen, J. M. Tiedje, and 
  A. P. Arkin. 2014. Stochasticity, succession, and environmental perturbations in 
  a fluidic ecosystem. Proceedings of the National Academy of Sciences of the United
  States of America 111:E836-E845.
}
\author{
Johon Quensen
}
