% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VectorSample.R
\name{VectorSample}
\alias{VectorSample}
\title{Randomize a binary vector}
\usage{
VectorSample(v, w)
}
\arguments{
\item{v}{A binary vector}

\item{w}{A vector of sampling weights; length equal to length of v.}
}
\value{
A randomized binary vector.
}
\description{
This function is used by Gotelli's matrix randomization functions. It takes an input 
  binary vector and a weight vector of equal length and reassigns 1s randomly in 
  proportion to the weights.
}
\examples{
sam <- sample(c(1,0), 10, replace=TRUE)
wt <- rep(1, 10) # Equal sampling probablilities.
random.sam <- VectorSample(sam, wt)
}
\references{
Gotelli, N. J. 2000. Null model analysis of species co-occurrence patterns.
  Ecology 81:2606-2621.
}
\author{
Nicholas J. Gotelli. Help edited by John Quensen.
}
